proto/auth: add cipher data to the AuthContainer message
AuthContainer is the envolope for all messages comming from the controller
or sent by EVE. For the LOC the node config should be encrypted. The best
place to provide encryption is to modify the `AuthContainer` itself, so that
the message can be both signed and encrypted (in this exact order).

This patch adds the `cipherData` and `cipherContext` members to the
`AuthContainer` to encrypt the payload. The decrypt and validate
procedure is the following:

  1. if `cipherData` is valid the encrypted payload is defined in the
     `cipherData.cipherData` member, the `protectedPayload` is set to null.
  2. Once EVE stack decrypts the payload, `protectedPayload` member is set
     with the decrypted data.
  3. The signature verification procedure can follow.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>