devmodelcommon: mark field 8 as reserved
Field 8 in the DeviceModelCommon message is marked as reserved to
prevent it from being used in the future through the protobuf
compiler.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>

vm: add flag to enable VNC for shim VM

We differentiate between the access to the edge application and the
shim VM:
- enableVNC is used to enable VNC for the edge application
- enable_vnc_shim_vm is used to enable VNC for the shim VM

This differentiation is need, because currently shim VM doesn't support
any user authentication, so anybody able to access the VNC port can
access the shim VM. This way VNC access to the shim VM stays disabled
by default unless explicitly enabled by the controller.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>

vm: convert enable VNC flag to snake_case

Signed-off-by: Paul Gaiduk <paulg@zededa.com>