Report MTU used by network instance
Now that network instance MTU is configurable, we should report what
the actual value being used is. This can differ from the configured
value in case it is invalid or conflicts with the MTU of the associated
device port.
Please note that we already report MTU of every device port.

Signed-off-by: Milan Lenco <milan@zededa.com>