Extend CellularAccessPoint with LTE attach configuration
LTE connection consists of two IP bearers, the initial EPS bearer
and the default EPS bearer. Device must first establish the initial
bearer (LTE attach procedure, which ModemManager shows as a transition
from the "enabled/searching" to "registered" modem state) and then it
connects to a default bearer (transition from "registered" to "connected").
Both bearers require PDP (Packet Data Protocol) context settings (APN,
ip-type, potentially username/password, etc.). Settings for the initial
bearer are typically provided by the modem profiles or the network, while
settings for the default bearer are user-configured. However, there are
cases, especially with LTE private networks or when changing SIM card
and moving to another operator for which the modem was not pre-configured,
where the modem may not provide the correct APN settings for the attach
procedure and it will fail to register into the network.
For these cases, it is usefull to have the LTE attach parameters
user-configurable.

Additionally, we will now allow the IP-type to be configured for both
bearers. Currently, EVE sequentially attempts all IP types until
a connection is established, causing unnecessary delays. Specifying
the correct IP type in the configuration will streamline the process.

Signed-off-by: Milan Lenco <milan@zededa.com>