Add Support for Single-Stack IP Networks
In EVE versions prior to 14, all networks operate in dual-stack mode,
regardless of whether the NetworkType is set to V4 or V6. Starting with
EVE 14 LTS, it is now possible to explicitly configure single-stack IP
networks by selecting V4Only or V6Only as the NetworkType.

However, if these single-stack options are used with older EVE versions,
the network will be marked as invalid and will not be applied. To prevent
this, the cloud should use the APICapability field to detect whether
single-stack support is available. If not, it should automatically revert
V4Only and V6Only to V4 and V6, respectively.

Signed-off-by: Milan Lenco <milan@zededa.com>